\name{CheckConsumptionData}
\alias{CheckConsumptionData}
\title{CheckConsumptionData}
\usage{
cons_data = CheckConsumptionData(pts, chem, cons, fill_gaps = FALSE)}
\arguments{
\item{pts}{...}
\item{chem}{...}
\item{cons}{...}
\item{fill_gaps}{...}
}
\value{cons_data}
\description{Temporary: runs Check_cons_v2, is able to fill potential gaps in the future.}
